% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hashtags.R
\name{extract_hashtags}
\alias{extract_hashtags}
\title{Extract hashtags from a vector of tweets}
\usage{
extract_hashtags(text)
}
\arguments{
\item{text}{a character vector of tweet texts}
}
\description{
Extract hashtags from a vector of tweets
}

