% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_urls.R
\name{extract_urls}
\alias{extract_urls}
\title{Extract URLs from a text file that contains database query results (such as Proquest)}
\usage{
extract_urls(file)
}
\arguments{
\item{file}{a character vector that specifies the filename}
}
\description{
Extract URLs from a text file that contains database query results (such as Proquest)
}

