% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tx.R
\name{flag_tx}
\alias{flag_tx}
\title{Flag transcript (a character vector) for whether it contains a key phrase}
\usage{
flag_tx(tx, filter_vector = "All Rights Reserved")
}
\arguments{
\item{tx}{a transcript or article that is a character vector}

\item{filter_vector}{a character vector containing key phrase(s)}
}
\description{
Flag transcript (a character vector) for whether it contains a key phrase
}

