% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dates.R
\name{is_blog}
\alias{is_blog}
\title{Determine whether a given article (from lexis nexis) is a blog post}
\usage{
is_blog(x, patt = "PUBLICATION-TYPE: Web Blog")
}
\arguments{
\item{x}{a character vector containing text for a single article}

\item{patt}{pattern that says that an article is a blog}
}
\description{
Determine whether a given article (from lexis nexis) is a blog post
}

