% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_date_row.R
\name{is_date}
\alias{is_date}
\title{Returns a logical to indicate whether input is a date, per a specific format}
\usage{
is_date(x, format = "\%B \%d, \%Y")
}
\arguments{
\item{x}{an input vector}

\item{format}{a character vector of length one specifying the date format}
}
\value{
a logical vector
}
\description{
Returns a logical to indicate whether input is a date, per a specific format
}

