% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_dates.R
\name{scrape_dates}
\alias{scrape_dates}
\title{Scrape webpage & extract dates for online media}
\usage{
scrape_dates(urls, html_nodes_argument = ".timestamp__date--published")
}
\arguments{
\item{urls}{a character vector of URLs}

\item{html_nodes_argument}{a character vector denoting the argument to html_nodes() function}
}
\description{
Scrape webpage & extract dates for online media
}

