% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_dates.R
\name{scrape_node}
\alias{scrape_node}
\title{Scrape webpage & extract a single node for online media}
\usage{
scrape_node(urls, node = ".timestamp__date--published")
}
\arguments{
\item{urls}{a character vector of URLs}

\item{node}{a character vector denoting the argument to html_nodes() function}
}
\description{
Scrape webpage & extract a single node for online media
}

