% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tx.R
\name{split_tx}
\alias{split_tx}
\title{Partition a text file, containing multiple articles or transcripts, into distinct articles}
\usage{
split_tx(tx, pattern)
}
\arguments{
\item{tx}{a character vector resulting from scan() of a text file containing multiple articles}

\item{pattern}{a character vector that is the pattern denoting the end of each article/story}
}
\value{
a list with length equal to the number of distinct articles
}
\description{
Partition a text file, containing multiple articles or transcripts, into distinct articles
}
