% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tx.R
\name{split_tx_trim}
\alias{split_tx_trim}
\title{Wrapper to partition text file into multiple articles AND to remove unneeded text at top & bottom of each article}
\usage{
split_tx_trim(tx, patt, headskip = 0, tailskip = 0)
}
\arguments{
\item{tx}{a character vector, resulting from reading, one entry per line, from a text file}

\item{patt}{pattern denoting start of a new article}

\item{headskip}{number of lines at top of each article to remove}

\item{tailskip}{number of lines at bottom of each article to remove}
}
\description{
Wrapper to partition text file into multiple articles AND to remove unneeded text at top & bottom of each article
}

