% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_date_row.R
\name{tabulate_dates}
\alias{tabulate_dates}
\title{Tabulate a vector of dates}
\usage{
tabulate_dates(directory, end_of_article = "All Rights Reserved")
}
\arguments{
\item{directory}{a character vector specifying path to a directory}

\item{end_of_article}{a character vector specifying end of story}
}
\description{
Tabulate a vector of dates
}
