% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dates.R
\name{trim}
\alias{trim}
\title{Remove extra white spaces before and after text.}
\usage{
trim(x)
}
\arguments{
\item{x}{a character vector to be 'trimmed'}
}
\value{
a character vector without preceding and trailing whitespace
}
\description{
Remove extra white spaces before and after text.
}
