% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tx.R
\name{trim_pattern}
\alias{trim_pattern}
\title{Find a pattern in an element of a character vector to identify where to start trimming & trim}
\usage{
trim_pattern(string, pattern, end = length(string))
}
\arguments{
\item{string}{the character vector in which to look for pattern}

\item{pattern}{the pattern to find in the strings}

\item{end}{latter index for subset to be removed}
}
\description{
Find a pattern in an element of a character vector to identify where to start trimming & trim
}

