% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_patterns.R
\name{trim_patterns}
\alias{trim_patterns}
\title{Find a pattern in an element of a character vector to identify where to start trimming & trim}
\usage{
trim_patterns(string, headpattern, tailpattern)
}
\arguments{
\item{string}{the character vector in which to look for pattern}

\item{headpattern}{the pattern to find in the strings}

\item{tailpattern}{the pattern to find in the strings}
}
\description{
Find a pattern in an element of a character vector to identify where to start trimming & trim
}
