% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_date_row.R
\name{find_date_row}
\alias{find_date_row}
\title{Find the row, for one article, in which the date is listed}
\usage{
find_date_row(x, format = "\%B \%d, \%Y")
}
\arguments{
\item{x}{a vector}
}
\value{
a number, the index of the first entry of x that is a date
}
\description{
Find the row, for one article, in which the date is listed
}

