% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tx.R
\name{trim_tx}
\alias{trim_tx}
\title{Remove unneeded text at top & bottom of each article}
\usage{
trim_tx(tx, headskip = 0, tailskip = 0)
}
\arguments{
\item{tx}{a character vector, resulting from reading, one entry per line, from a text file}

\item{headskip}{number of lines at top of each article to remove}

\item{tailskip}{number of lines at bottom of each article to remove}
}
\description{
Remove unneeded text at top & bottom of each article
}

